Bug - Device index gets remapped
Instead get all device info from pyaudio and search the new index
by device_name.
