+++ README.md
@@ -1,4 +1,9 @@
 # Setup
+## For Linux only
+Install portaudio package, windows will install this with package with pyaudio.  
+`sudo apt-get update && sudo apt-get install portaudio19-dev`  
+
+## Continued setup.
 Run `pip install -r requirements.txt`  
 Run `python setup.py` to create the data directories and populate JOB_DEVICE_MAPPINGS  
 See below for steps to generate the summaries locally vs with OpenAI  
@@ -36,6 +41,12 @@ Choose between the models below as specified here https://github.com/openai/whis
 tiny, tiny.en, base, base.en, small, small.en, medium, medium.en, large, N/A  
 Whisper models are saved in "~/.cache/whisper"  
 
+## Capturing Desktop Audio
+pyaudio can only read input audio streams, so there's no built-in way to capture desktop audio. However you can install any audio loopback tools to route your desktop audio to a virtual microphone. See some software recommendations below.  
+Linux - PulseAudio, ALSA, JACK  
+MacOS - Loopback, VB Audio, JACK  
+Windows - Stereo Mix (by Realtek, needs to be enabled)  
+
 # Usage
 Run `python main.py <job_index> <meeting_name>`  
 Steps:  
@@ -47,4 +58,4 @@ Steps:
 
 # Todo
 - Feature: In job device mappings, use an index (ie 1, 2) or the company name (ie J1, J2) to trigger the main script  
-- Bug: pyaudio device indexes will sometimes get remapped, forcing the user to re-run setup.py. The device name however shouldn't change unless you change the port. Specific to the soundcraft mixer, the prefix name of these devices (ex Mic/Line In 11/12) never changes - make a separate branch to support this add'tl feature.  
+- Bug: pyaudio device indexes will sometimes get remapped, forcing the user to re-run setup.py. The device name however shouldn't change unless you change the port. Specific to the soundcraft mixer, the prefix name of these devices (ex Mic/Line In 11/12) never changes - make a separate branch to support this add'tl feature. - Status: Currently in dev branch
